RestaurantRestController.PATCH

+++ src/main/java/rest/DishOfTheDay/service/RestaurantService.java
@@ -11,8 +11,6 @@ import rest.DishOfTheDay.util.exception.NotFoundException;
 import java.util.List;
 import java.util.Optional;
 
-import static rest.DishOfTheDay.util.ValidationUtil.checkNotFoundWithId;
-
 @Service
 public class RestaurantService {
 
@@ -27,23 +25,23 @@ public class RestaurantService {
         return repository.findAll(new Sort(Sort.Direction.ASC, "name"));
     }
 
-    public Optional<Restaurant> get(Integer id) {
-        return repository.findById(id);
-        //return checkNotFoundWithId(repository.findById(id).get(), id, Restaurant.class);
+    public Restaurant get(Integer id) {
+        Optional<Restaurant> restaurant = repository.findById(id);
+        if(restaurant.isPresent())
+            return restaurant.get();
+        else
+            throw new NotFoundException(Restaurant.class);
     }
 
-    public Restaurant create(Restaurant restaurant) {
+    public Restaurant save(Restaurant restaurant) {
         Assert.notNull(restaurant, "restaurant must not be null");
         return repository.save(restaurant);
     }
 
-    public void delete (int id) throws NotFoundException {
-        checkNotFoundWithId(repository.existsById(id), id, Restaurant.class);
-        repository.deleteById(id);
-    }
-
-    public void update(Restaurant restaurant) {
-        Assert.notNull(restaurant, "restaurant must not be null");
-        repository.save(restaurant);
+    public void delete (int id) {
+        if (repository.existsById(id))
+            repository.deleteById(id);
+        else
+            throw new NotFoundException(Restaurant.class);
     }
 }
